requirements.txt
APScheduler==3.10.4
async-timeout==4.0.3
blinker==1.8.2
cachelib==0.9.0
click==8.1.7
Deprecated==1.2.14
dnspython==2.6.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Limiter==3.8.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
Pygments==2.18.0
pymongo==4.10.0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.0
rich==13.9.1
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.4
wrapt==1.16.0
